+++ Assets/Scenes/LoadingScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fae80038dbe41417e90537dd9b5923f4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
