+++ Assets/Scenes/StarMap.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 18f49deb221564229a32c64679fc9e83
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
