+++ Assets/Scripts/Managers/StarMap.cs.meta
