+++ ProjectSettings/EditorBuildSettings.asset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:aa2b572bb928058de1da32bb7d7423e8ff6996112231b538f98a7c3fede181a9
-size 261
+oid sha256:c2910f553377a489fea82359584ba1a4eb7b27206b94376527edb0f6d340978f
+size 354
